+++ bactrack/gui/gui.py
@@ -1,68 +1,283 @@
-from PyQt6.QtWidgets import QApplication, QMainWindow, QVBoxLayout, QWidget, QLabel
-from PyQt6.QtCore import Qt
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.figure import Figure
-from viz import Viz 
+import sys
+
+
+from PyQt6.QtCore import Qt, QSize, QObject, QEvent, pyqtSignal, QThread
+from PyQt6.QtWidgets import (QApplication, QMainWindow, QWidget, QVBoxLayout, QPlainTextEdit,
+                             QHBoxLayout, QComboBox, QPushButton, QTextEdit, QScrollBar, QLabel, QStackedWidget, QSplitter)
+
+
+import threading
 from run import run_track
+from viz import Viz 
+import logging
+from lineage import Lineage
+from visualizer import get_graph_stats_text
 
-# from your_viz_module import Viz
-# from your_tracking_module import run_track
+        
+class StreamRedirect(QObject):
+    text_written = pyqtSignal(str)
+    
+    def write(self, text):
+        self.text_written.emit(str(text))
+    
+    def flush(self):
+        pass
 
-class MainWindow(QMainWindow):
+
+class Worker(QObject):
+    finished = pyqtSignal(object, object)  # Signal for when the task is finished
+    error = pyqtSignal(Exception)  # Signal for emitting an error
+
+    def __init__(self, folder_path):
+        super().__init__()
+        self.folder_path = folder_path
+
+    def run_track(self):
+        try:
+            # Placeholder for the time-consuming task
+            composer, G = run_track(self.folder_path)
+            self.finished.emit(composer, G)  # Emit the results upon completion
+        except Exception as e:
+            self.error.emit(e)  # Emit any errors encountered during the task
+
+class QSignalEmitter(QObject):
+    # Define a signal to emit log messages
+    emit_log = pyqtSignal(str)
+
+class QTextEditLogger(logging.Handler, QSignalEmitter):
+    def __init__(self):
+        logging.Handler.__init__(self)
+        QSignalEmitter.__init__(self)
+
+    def emit(self, record):
+        message = self.format(record)
+        self.emit_log.emit(message)
+    
+
+class ToggleButton(QPushButton):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setCheckable(True)
+        self.setChecked(False)
+        self.setText("Raw Image Off")
+    
+
+class MyMainWindow(QMainWindow):
     def __init__(self):
         super().__init__()
-        self.setWindowTitle("Backtrack")
-        self.setGeometry(100, 100, 800, 600)  # Adjust size as needed
 
-        self.main_widget = QWidget(self)
+        self.frame = 0 
+        self.label_index = 0
+        self.style_index = 0
+
+        # Create main widget and layout
+        self.main_widget = QWidget()
+        self.main_layout = QHBoxLayout(self.main_widget)
+        self.setGeometry(100, 100, 800, 600)
+
+        # Redirect stdout
+        self.stream_redirect = StreamRedirect()
+        self.stream_redirect.text_written.connect(self.log_to_terminal)
+        sys.stdout = self.stream_redirect
+
+        # Redirect logging
+        log_handler = QTextEditLogger()
+        log_handler.emit_log.connect(self.log_to_terminal)
+        logging.basicConfig(level=logging.DEBUG, handlers=[log_handler])
+
+
+        # Preparing all wigets
+        self.type_dropdown = QComboBox()
+        self.model_dropdown = QComboBox()
+        self.run_button = QPushButton('Run Tracking')
+        self.reset_button = QPushButton('Reset zoom')
+        self.toggle_button = ToggleButton(self)
+        self.scrollbar = QScrollBar(Qt.Orientation.Horizontal)
+
+        self.main_canvas = QStackedWidget()
+        self.track_timelapse_canvas = Viz(self)
+
+        self.lineage = Lineage()
+
+
+        #  ================== Left vertical section ================
+        left_layout = QVBoxLayout()
+        # Create the left section with the button and dropdowns
+            # Type drop down 
+        self.type_dropdown.addItems([".png", ".tif", ".jpg"])
+        type_layout = QHBoxLayout()  
+        type_layout.addWidget( QLabel('Type:') , 1)
+        type_layout.addWidget(self.type_dropdown, 3)
+
+            # Model drop down
+        self.model_dropdown.addItems(["Model 1", "Model 2", "Model 3"])
+        model_layout = QHBoxLayout()  
+        model_layout.addWidget(QLabel('model:') , 1)
+        model_layout.addWidget(self.model_dropdown, 3)
+
+            # Buttons
+        self.toggle_button.toggled.connect(self.rawImageToggled)
+        self.reset_button.clicked.connect(self.track_timelapse_canvas.reset_zoom)
+
+        # Left layout construction
+        left_layout.addLayout(type_layout)
+        left_layout.addLayout(model_layout)
+        left_layout.addWidget(self.toggle_button)
+        left_layout.addWidget(self.reset_button)
+        left_layout.addWidget(self.run_button)
+        left_layout.addStretch(1) 
+
+        #  ================== Right vertical section ================
+        right_layout = QVBoxLayout()
+        # Create the right section with the canvas and terminal
+            # Main canvas section:
+        #self.main_canvas.setFixedHeight(400)
+                # Text placeholder
+        self.text_widget = QLabel("Drop Your Folder Here")
+        self.text_widget.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        self.text_widget.setStyleSheet("border: 2px dashed #000000;")
+        self.main_canvas.addWidget(self.text_widget)
+                # Video canvas placeholder
+        self.main_canvas.addWidget(self.track_timelapse_canvas)
+        self.main_canvas.setCurrentIndex(0)
+        self.main_canvas.setFocus()
+
+                # Scrollbar section:
+        self.scrollbar.valueChanged.connect(self.updateFrameView)
+
+                # Terminal section:
+        self.terminal = QPlainTextEdit(self)
+        self.terminal.setReadOnly(True)
+
+        
+        # Right layout construction
+        right_layout.addWidget(self.main_canvas, 5)
+        right_layout.addWidget(self.scrollbar, 1) # Placeholder for the scrollbar
+        right_layout.addWidget(self.terminal, 1)
+
+
+        #  ================== Option section: Linage ================
+        self.linage_stat = QLabel("Linage")
+        self.linage_stat.setFixedHeight(100)
+        extra_layout = QVBoxLayout()
+        extra_layout.addWidget(self.linage_stat)
+        extra_layout.addWidget(self.lineage)
+
+
+        # Create the splitter and add the two layouts
+        splitter = QSplitter(Qt.Orientation.Horizontal)
+        left_widget = QWidget()
+        left_widget.setLayout(left_layout)
+        right_widget = QWidget()
+        right_widget.setLayout(right_layout)
+
+        extra_widget = QWidget()
+        extra_widget.setLayout(extra_layout)
+
+        # Finalize everything 
+        splitter.addWidget(left_widget)
+        splitter.addWidget(right_widget)
+        splitter.addWidget(extra_widget)
+
+        extra_widget.setVisible(False)
+        self.extra_widget = extra_widget
+
+        splitter.setSizes([200, 600, 300])
+        
+       
+        # ================== Put everthing together main section setup ================
+        # Set the splitter as the central widget
+        self.main_layout.addWidget(splitter)
         self.setCentralWidget(self.main_widget)
-        self.layout = QVBoxLayout(self.main_widget)
 
-        self.drag_label = QLabel("Drag Folder Here")
-        self.drag_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        self.layout.addWidget(self.drag_label)
+        # Set window title and show the application
+        self.setWindowTitle("Bactrack")
+        self.show()
 
-        self.figure = Figure()
-        self.canvas = FigureCanvas(self.figure)
-        self.layout.addWidget(self.canvas)
-        self.viz = None
+    def log_to_terminal(self, message):
+        self.terminal.insertPlainText(message)
+        self.terminal.ensureCursorVisible()
+    
 
-
-    def init_viz(self, composer, G):
-        # Remove placeholder label if it exists
-        if self.drag_label:
-            self.drag_label.deleteLater()
-            self.drag_label = None
-
-        # Initialize Viz with the existing figure
-        if self.viz is None:
-            self.viz = Viz(composer, G, self.figure)
+    def rawImageToggled(self, checked):
+        if checked:
+            self.toggle_button.setText("Raw image On")
+            self.main_canvas.setCurrentIndex(1)
         else:
-            # If viz already exists, just update it
-            self.viz.update_plot()
-
-        self.canvas.show()
-        self.canvas.draw_idle()
-
+            self.toggle_button.setText("Raw image Off")
+            self.main_canvas.setCurrentIndex(0)
 
     def dropEvent(self, event):
         urls = event.mimeData().urls()
+        print(f"{urls} dropped")
         if urls:
             folder_path = urls[0].toLocalFile()
-            # You need to define how you obtain `composer` and `G` from the folder_path
-            # This might involve calling a function similar to run_track(folder_path)
-            # Assuming run_track returns `composer` and `G`
-            composer, G = run_track(folder_path)
-            self.init_viz(composer, G)
-            self.canvas.draw_idle() 
+
+            # Setup the worker and thread
+            self.thread = QThread()  # Create a QThread object
+            self.worker = Worker(folder_path)  # Create a Worker object
+            self.worker.moveToThread(self.thread)  # Move the Worker to the thread
+
+            # Connect signals
+            self.worker.finished.connect(self.on_track_finished)  # Connect the finished signal to a slot
+            self.worker.finished.connect(self.thread.quit)  # Ensure the thread quits when the task is done
+            self.worker.error.connect(self.on_track_error)  # Connect the error signal to a slot
+            self.thread.started.connect(self.worker.run_track)  # Start the worker's task when the thread starts
+
+            # Start the thread
+            self.thread.start()
+
+            # Optional: Change the cursor to indicate processing
+            QApplication.setOverrideCursor(Qt.BusyCursor)
+
 
     def dragEnterEvent(self, event):
         if event.mimeData().hasUrls():
             event.acceptProposedAction()
 
+    def on_track_finished(self, composer, G):
+        QApplication.restoreOverrideCursor()  # Restore the cursor
+        self.track_timelapse_canvas.run(composer, G)
+        self.main_canvas.setCurrentIndex(1)
+        self.scrollbar.setMaximum(self.track_timelapse_canvas.max_frame)
+        self.lineage.show(G)
+        self.linage_stat.setText(get_graph_stats_text(G))
+        self.setGeometry(100, 100, 1100, 600)
+        self.extra_widget.setVisible(True)
 
-if __name__ == "__main__":
-    app = QApplication([])
-    window = MainWindow()
-    window.show()
-    app.exec()
+    def on_track_error(self, error):
+        QApplication.restoreOverrideCursor()  # Restore the cursor
+        print(f"Error: {error}")  # Handle the error (e.g., show a message box)
+
+    def updateFrameView(self, value):
+        self.frame = value
+        self.track_timelapse_canvas.update_plot(self)
+    
+    def keyPressEvent(self, event):
+        if self.main_canvas.currentIndex() != 1:
+            event.ignore()
+            return
+        if event.key() == Qt.Key_BracketRight:
+            self.frame = min(self.frame + 1, self.track_timelapse_canvas.max_frame) 
+        elif event.key() == Qt.Key_BracketLeft:
+            self.frame = max(self.frame - 1, 0)
+        elif event.key() == Qt.Key_C:
+            self.style_index = (self.style_index + 1) % len(self.track_timelapse_canvas.label_styles)
+        elif event.key() == Qt.Key_L:
+            self.label_index = (self.label_index + 1) % len(self.track_timelapse_canvas.labels)
+        else:
+            event.ignore() 
+            return  
+        
+        self.track_timelapse_canvas.update_plot(self)
+
+
+def main():
+    app = QApplication(sys.argv)
+    mainWindow = MyMainWindow()
+    mainWindow.log_to_terminal("Bactrack start.")
+    mainWindow.show()
+    sys.exit(app.exec())
+
+if __name__ == '__main__':
+    main()
